requirements.txt
amqp==2.6.1
asgiref==3.5.0
attrs==21.4.0
billiard==3.6.4.0
celery==4.4.7
certifi==2021.10.8
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
coverage==6.3.2
Django==3.2.2
django-cors-headers==3.11.0
django-debug-toolbar==3.2.4
django-extensions==3.1.5
django-filter==21.1
django-js-asset==2.0.0
django-mptt==0.13.4
django-oauth-toolkit==1.3.3
django-oauth2-provider==0.2.6.1
django-rest-swagger==2.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==4.4.0
drf-yasg==1.20.0
environs==9.5.0
flower==0.9.7
gunicorn==20.0.0
humanize==4.0.0
idna==3.3
inflection==0.5.1
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.1.1
kombu==4.6.11
MarkupSafe==2.1.1
marshmallow==3.15.0
oauthlib==3.2.0
openapi-codec==1.3.2
packaging==21.3
Pillow==9.0.1
pluggy==1.0.0
prometheus-client==0.8.0
psycopg2-binary==2.8.6
py==1.11.0
PyJWT==2.3.0
pyparsing==3.0.7
pytest==7.1.1
pytest-cov==3.0.0
pytest-django==4.5.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
shortuuid==1.0.8
simplejson==3.17.6
sqlparse==0.4.2
tomli==2.0.1
tornado==6.1
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
vine==1.3.0
